Rename base module from What2cook to What2Cook
This way the capitalization is consistent.
It was necessary to create an alias in mix.exs, since otherwise
phx.routes would not work without extra args.
Funnily enough a dummy app with the right arguments given to phx.new
works just fine... I might transition to that if the problems persist
after a clean git clone, given that the diff did not indicate any
meaningful difference.